IKASAN-150 - First 1/2 of projects done

+++ framework/src/main/java/org/ikasan/framework/management/search/PagedSearchResult.java
@@ -76,5 +76,5 @@ public interface PagedSearchResult<T> extends List<T> {
 	 * 
 	 * @return size of the larger super result set
 	 */
-	public int getResultSize();
+	public long getResultSize();
 }
